Made the change to Jo's Login Screen
